Solve circular dependencies, isolate ROMS blueprint from core, add recursive app auto-import
