+++ cstar/applications/hello_world_app.py
@@ -1,9 +1,16 @@
 import typing as t
 
-from cstar.applications.utils import register_application
+from cstar.applications.utils import (
+    ApplicationDefinition,
+    Transform,
+    register_application,
+)
 from cstar.entrypoint.xrunner import XBlueprintRunner, XRunnerResult
 from cstar.execution.handler import ExecutionStatus
-from cstar.orchestration.models import ApplicationDefinition, Blueprint, Transform
+from cstar.orchestration.models import Blueprint
+
+if t.TYPE_CHECKING:
+    from cstar.orchestration.orchestration import LiveStep
 
 APP_NAME: t.Literal["hello_world"] = "hello_world"
 
@@ -50,4 +57,4 @@ class HelloWorldApplication(
     long_name = APP_NAME
     runner = HelloWorldRunner
     blueprint = HelloWorldBlueprint
-    applicable_transforms: tuple[type[Transform], ...] = ()
+    applicable_transforms: "tuple[type[Transform[LiveStep]], ...]" = ()
