+++ cstar/orchestration/dag_runner.py
@@ -8,6 +8,7 @@ from pathlib import Path
 
 from prefect import flow
 
+from cstar.applications.roms_marbl.transforms import RomsMarblTimeSplitter
 from cstar.base.env import capture_environment
 from cstar.base.log import get_logger
 from cstar.execution.file_system import DirectoryManager
@@ -27,7 +28,6 @@ from cstar.orchestration.serialization import deserialize, serialize
 from cstar.orchestration.state import load_sentinels
 from cstar.orchestration.tracking import TrackingRepository, WorkplanRun
 from cstar.orchestration.transforms import (
-    RomsMarblTimeSplitter,
     TemplateFillTransform,
     WorkplanTransformer,
 )
